pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aula-ai-tutor"
version = "0.1.0"
description = "Sistema de agentes educativos con ADK y A2A para tutoría socrática"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "ThaleOn AI Systems", email = "contact@thaleon.ai"}
]
keywords = ["ai", "education", "agents", "tutoring", "socratic-method"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Education",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Google ADK
    "google-adk>=0.1.0",
    
    # Model backends
    "ollama>=0.4.0",
    "openai>=1.50.0",
    "transformers>=4.45.0",
    "torch>=2.2.0",
    "accelerate>=0.30.0",
    "bitsandbytes>=0.43.0",
    
    # A2A Protocol
    "httpx>=0.27.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "websockets>=13.0",
    
    # Utilities
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "pyyaml>=6.0.0",
    "jinja2>=3.1.0",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
    
    # Text processing for guardrails
    "sentence-transformers>=3.2.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.5.0",
    "rapidfuzz>=3.10.0",
    
    # Async support
    "anyio>=4.6.0",
    "aiofiles>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pre-commit>=3.8.0",
    "httpx>=0.27.0",
]

ui = [
    "gradio>=5.0.0",
    "streamlit>=1.38.0",
]

[project.scripts]
aula-solver = "scripts.start_solver:main"
aula-tutor = "scripts.start_tutor:main"
aula-demo = "scripts.demo_session:main"

[project.urls]
Homepage = "https://github.com/thaleon/aula-ai-tutor"
Documentation = "https://github.com/thaleon/aula-ai-tutor#readme"
Repository = "https://github.com/thaleon/aula-ai-tutor"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing"
filterwarnings = [
    "ignore::DeprecationWarning",
]
